style(windows): silence MSVC C4305/C4244 conversion warnings in F2Dock sources (#14)
Mechanically apply C-style casts to silence MSVC double-to-float (C4305)
and double-to-int / size_t conversion warnings (C4244) reported by the
Windows Debug CI job. Casts use the existing F2Dock idiom `(T)(expr)`,
matching patterns already present in the codebase (e.g. `(int)ceil(...)`).

Changes:
- inc/f2dock/ElementInformation.h: append `f` suffix to float-field
  literals in the element-property table (310 sites).
- src/f2dock/Docking.cpp: wrap `ceil(...)`/`floor(...)` int-target
  initializers and Matrix::translation / Vector / retrieveTransformation
  arguments (110 sites).
- src/fft-utils/fastfft.cpp: cast `sizeof(...)*N` to `size_t` for
  FFTW_malloc, cast grid dims to int (33 sites).
- src/PostFiltering/FilterOutput.cpp (16), src/fast-clash/clashFilter.cpp
  (9), src/fast-PQ/PairingHeap.cpp (9), src/fast-resCont/resContFilter.cpp
  (9), src/GB-rerank/pairingHeap.cpp (9), src/f2dock/F2Dock.cpp (9),
  src/math/Matrix.cpp (2), src/f2dock/TopValues.cpp (1).

Of the ~678 warning sites in F2Dock-owned code, 517 are addressed by
this scripted pass. The remaining ~161 sites are skipped — they involve
multi-line expressions, dependent-type STL container `.size()` returns,
SSE intrinsic helpers, and complex PG-range arithmetic that need manual
review. They will be tackled separately.

libcvc warnings (`build/_deps/libcvc-src/...`) are deliberately not
touched in this PR and will be addressed in libcvc upstream.

Linux/macOS GCC and Clang do not emit C4305/C4244, so no functional
change is expected on those platforms; the casts are equivalent to the
implicit conversions they replace.

+++ src/fft-utils/fastfft.cpp
@@ -95,9 +95,9 @@ void gridding(int M, double *x, double *y, double *z, float *r, char *type,
     // influence; observe that xL, xU, yL, yU, zL, zU \in [ -n/2, n/2 )
     int xL, xU, yL, yU, zL, zU;
 
-    xL = floor(xc) - m, xU = ceil(xc) + m;
-    yL = floor(yc) - m, yU = ceil(yc) + m;
-    zL = floor(zc) - m, zU = ceil(zc) + m;
+    xL = floor(xc) - m, xU = (int)(ceil(xc) + m);
+    yL = floor(yc) - m, yU = (int)(ceil(yc) + m);
+    zL = floor(zc) - m, zU = (int)(ceil(zc) + m);
 
     // apply atom s's influence
     for (int zt = zL; zt <= zU; zt++)
@@ -172,9 +172,9 @@ void gridding(int M, double *x, double *y, double *z, float *r, char *type,
 
         // compute the region [ xL, xU ] x [ yL, yU ] x [ zL, zU ] of atom s's
         // influence; observe that xL, xU, yL, yU, zL, zU \in [ -n/2, n/2 )
-        int xL = floor(xc) - md, xU = ceil(xc) + md;
-        int yL = floor(yc) - md, yU = ceil(yc) + md;
-        int zL = floor(zc) - md, zU = ceil(zc) + md;
+        int xL = (int)(floor(xc) - md), xU = (int)(ceil(xc) + md);
+        int yL = (int)(floor(yc) - md), yU = (int)(ceil(yc) + md);
+        int zL = (int)(floor(zc) - md), zU = (int)(ceil(zc) + md);
 
         if (xL < -n / 2)
           xL = -n / 2;
@@ -266,9 +266,9 @@ void griddingElec(int M, double *x, double *y, double *z, float *r, char *type,
 
     // compute the region [ xL, xU ] x [ yL, yU ] x [ zL, zU ] of atom s's
     // influence; observe that xL, xU, yL, yU, zL, zU \in [ -n/2, n/2 )
-    int xL = floor(xc) - rc, xU = ceil(xc) + rc;
-    int yL = floor(yc) - rc, yU = ceil(yc) + rc;
-    int zL = floor(zc) - rc, zU = ceil(zc) + rc;
+    int xL = (int)(floor(xc) - rc), xU = (int)(ceil(xc) + rc);
+    int yL = (int)(floor(yc) - rc), yU = (int)(ceil(yc) + rc);
+    int zL = (int)(floor(zc) - rc), zU = (int)(ceil(zc) + rc);
 
     if (xL < -(n >> 1))
       xL = -(n >> 1);
@@ -328,9 +328,9 @@ void griddingHbond(int M, double *x, double *y, double *z, float *r,
 
     // compute the region [ xL, xU ] x [ yL, yU ] x [ zL, zU ] of atom s's
     // influence; observe that xL, xU, yL, yU, zL, zU \in [ -n/2, n/2 )
-    int xL = floor(xc) - rc, xU = ceil(xc) + rc;
-    int yL = floor(yc) - rc, yU = ceil(yc) + rc;
-    int zL = floor(zc) - rc, zU = ceil(zc) + rc;
+    int xL = (int)(floor(xc) - rc), xU = (int)(ceil(xc) + rc);
+    int yL = (int)(floor(yc) - rc), yU = (int)(ceil(yc) + rc);
+    int zL = (int)(floor(zc) - rc), zU = (int)(ceil(zc) + rc);
 
     if (xL < -(n >> 1))
       xL = -(n >> 1);
@@ -386,9 +386,9 @@ void griddingHydrophobicity(int M, double *x, double *y, double *z, float *r,
 
     // compute the region [ xL, xU ] x [ yL, yU ] x [ zL, zU ] of atom s's
     // influence; observe that xL, xU, yL, yU, zL, zU \in [ -n/2, n/2 )
-    int xL = floor(xc) - rc, xU = ceil(xc) + rc;
-    int yL = floor(yc) - rc, yU = ceil(yc) + rc;
-    int zL = floor(zc) - rc, zU = ceil(zc) + rc;
+    int xL = (int)(floor(xc) - rc), xU = (int)(ceil(xc) + rc);
+    int yL = (int)(floor(yc) - rc), yU = (int)(ceil(yc) + rc);
+    int zL = (int)(floor(zc) - rc), zU = (int)(ceil(zc) + rc);
 
     if (xL < -(n >> 1))
       xL = -(n >> 1);
@@ -454,9 +454,9 @@ void griddingSimpleComplementarity(int M, double *x, double *y, double *z,
 
     // compute the region [ xL, xU ] x [ yL, yU ] x [ zL, zU ] of atom s's
     // influence; observe that xL, xU, yL, yU, zL, zU \in [ -n/2, n/2 )
-    int xL = floor(xc) - rc, xU = ceil(xc) + rc;
-    int yL = floor(yc) - rc, yU = ceil(yc) + rc;
-    int zL = floor(zc) - rc, zU = ceil(zc) + rc;
+    int xL = (int)(floor(xc) - rc), xU = (int)(ceil(xc) + rc);
+    int yL = (int)(floor(yc) - rc), yU = (int)(ceil(yc) + rc);
+    int zL = (int)(floor(zc) - rc), zU = (int)(ceil(zc) + rc);
 
     if (xL < -(n >> 1))
       xL = -(n >> 1);
